refactor(settings): split SettingsWindow code-behind
