+++ Views/SettingsWindow.xaml.cs
@@ -1,468 +1,13 @@
 using System;
 using System.ComponentModel;
-using System.Diagnostics;
-using System.Runtime.CompilerServices;
 using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Media;
-using System.Windows.Input;
-using System.Windows.Media.Imaging;
-using System.Windows.Threading;
 using MaterialDesignThemes.Wpf;
-using Microsoft.Win32;
-using System.IO;
 using WindBoard.Services;
-using WindBoard.Core.Ink;
 
 namespace WindBoard
 {
     public partial class SettingsWindow : Window, INotifyPropertyChanged
     {
-        // 主窗口依赖已移除（仅保留 UI）
-        private Color _currentColor;
-        private PopupBox? _colorPopupBox;
-
-        // --- 视频展台设置（SettingsWindow 层） ---
-        private bool _videoPresenterEnabled;
-        private string _videoPresenterPath = string.Empty;
-        private string _videoPresenterArgs = string.Empty;
-
-        // --- 伪装设置 ---
-        private bool _camouflageEnabled;
-        private string _camouflageTitle = string.Empty;
-        private string _camouflageSourcePath = string.Empty;
-        private ImageSource? _camouflageIconPreview;
-        private string _camouflageSourceDisplayName = "未选择文件";
-
-        // --- 书写设置 ---
-        private bool _strokeThicknessConsistencyEnabled;
-        private bool _simulatedPressureEnabled;
-        private double _simulatedPressureStartTaperMm;
-        private double _simulatedPressureEndTaperMm;
-        private double _simulatedPressureSpeedMinMmPerSec;
-        private double _simulatedPressureSpeedMaxMmPerSec;
-        private double _simulatedPressureFastSpeedMinFactor;
-        private double _simulatedPressureFloor;
-        private double _simulatedPressureEndFloor;
-        private double _simulatedPressureSmoothingTauMs;
-        private DispatcherTimer? _simulatedPressurePersistTimer;
-        private bool _isApplyingSimulatedPressureFields;
-
-        private const string DefaultVideoPresenterPath = @"C:\\Program Files (x86)\\Seewo\\EasiCamera\\sweclauncher\\sweclauncher.exe";
-        private const string DefaultVideoPresenterArgs = "-from en5";
-
-        private static readonly SimulatedPressureConfig DefaultSimulatedPressureConfig = SimulatedPressureConfig.CreateDefault();
-
-        public event PropertyChangedEventHandler? PropertyChanged;
-
-        protected virtual void OnPropertyChanged([CallerMemberName] string? propertyName = null)
-        {
-            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
-        }
-
-        public Color CurrentColor
-        {
-            get => _currentColor;
-            set
-            {
-                if (_currentColor != value)
-                {
-                    _currentColor = value;
-                    OnPropertyChanged();
-                    OnPropertyChanged(nameof(CurrentColorHex));
-                    OnPropertyChanged(nameof(CurrentBrush));
-
-                    SettingsService.Instance.SetBackgroundColor(_currentColor);
-                }
-            }
-        }
-
-        public string CurrentColorHex
-        {
-            get => _currentColor.ToString();
-            set
-            {
-                if (string.IsNullOrWhiteSpace(value)) return;
-                var hex = value.Trim();
-                if (!hex.StartsWith("#")) hex = "#" + hex;
-                try
-                {
-                    var color = (Color)ColorConverter.ConvertFromString(hex);
-                    CurrentColor = color;
-                }
-                catch (Exception)
-                {
-                    // 无效颜色时默认使用白色
-                    CurrentColor = Colors.White;
-                }
-            }
-        }
-
-        // 供 UI 直接绑定刷子，避免 SolidColorBrush.Color 子属性绑定更新问题
-        public SolidColorBrush CurrentBrush => new SolidColorBrush(_currentColor);
-
-        // --- 视频展台属性（绑定到 XAML，通过 ElementName=SettingsWindowRoot） ---
-        public bool VideoPresenterEnabled
-        {
-            get => _videoPresenterEnabled;
-            set
-            {
-                if (_videoPresenterEnabled != value)
-                {
-                    _videoPresenterEnabled = value;
-                    OnPropertyChanged();
-                    // 立即持久化：与颜色设置一致（切换开关即时生效）
-                    try { SettingsService.Instance.SetVideoPresenterEnabled(value); } catch { }
-                }
-            }
-        }
-
-        public string VideoPresenterPath
-        {
-            get => _videoPresenterPath;
-            set
-            {
-                if (_videoPresenterPath != value)
-                {
-                    _videoPresenterPath = value ?? string.Empty;
-                    OnPropertyChanged();
-                }
-            }
-        }
-
-        public string VideoPresenterArgs
-        {
-            get => _videoPresenterArgs;
-            set
-            {
-                if (_videoPresenterArgs != value)
-                {
-                    _videoPresenterArgs = value ?? string.Empty;
-                    OnPropertyChanged();
-                }
-            }
-        }
-
-        // --- 伪装属性（绑定到 XAML，通过 ElementName=SettingsWindowRoot） ---
-        public bool CamouflageEnabled
-        {
-            get => _camouflageEnabled;
-            set
-            {
-                if (_camouflageEnabled != value)
-                {
-                    _camouflageEnabled = value;
-                    OnPropertyChanged();
-                    try { SettingsService.Instance.SetCamouflageEnabled(value); }
-                    catch (Exception ex) { Debug.WriteLine($"[Settings] Failed to persist camouflage enabled flag: {ex}"); }
-                }
-            }
-        }
-
-        public string CamouflageTitle
-        {
-            get => _camouflageTitle;
-            set
-            {
-                if (_camouflageTitle != value)
-                {
-                    _camouflageTitle = value ?? string.Empty;
-                    OnPropertyChanged();
-                    try { SettingsService.Instance.SetCamouflageTitle(_camouflageTitle); }
-                    catch (Exception ex) { Debug.WriteLine($"[Settings] Failed to persist camouflage title: {ex}"); }
-                }
-            }
-        }
-
-        public string CamouflageSourcePath
-        {
-            get => _camouflageSourcePath;
-            set
-            {
-                if (_camouflageSourcePath != value)
-                {
-                    _camouflageSourcePath = value ?? string.Empty;
-                    OnPropertyChanged();
-                    OnPropertyChanged(nameof(CamouflageSourceDisplayName));
-                    try { SettingsService.Instance.SetCamouflageSourcePath(_camouflageSourcePath); }
-                    catch (Exception ex) { Debug.WriteLine($"[Settings] Failed to persist camouflage source path: {ex}"); }
-                    RefreshCamouflagePreview(buildCache: true);
-                }
-            }
-        }
-
-        public ImageSource? CamouflageIconPreview
-        {
-            get => _camouflageIconPreview;
-            private set
-            {
-                if (_camouflageIconPreview != value)
-                {
-                    _camouflageIconPreview = value;
-                    OnPropertyChanged();
-                }
-            }
-        }
-
-        public string CamouflageSourceDisplayName
-        {
-            get => _camouflageSourceDisplayName;
-            private set
-            {
-                if (_camouflageSourceDisplayName != value)
-                {
-                    _camouflageSourceDisplayName = value;
-                    OnPropertyChanged();
-                }
-            }
-        }
-
-        public bool StrokeThicknessConsistencyEnabled
-        {
-            get => _strokeThicknessConsistencyEnabled;
-            set
-            {
-                if (_strokeThicknessConsistencyEnabled != value)
-                {
-                    _strokeThicknessConsistencyEnabled = value;
-                    OnPropertyChanged();
-                    try
-                    {
-                        SettingsService.Instance.SetStrokeThicknessConsistencyEnabled(value);
-                    }
-                    catch (Exception ex)
-                    {
-                        System.Diagnostics.Debug.WriteLine(
-                            $"Failed to persist StrokeThicknessConsistencyEnabled setting: {ex}");
-                    }
-                }
-            }
-        }
-
-        public bool SimulatedPressureEnabled
-        {
-            get => _simulatedPressureEnabled;
-            set
-            {
-                if (_simulatedPressureEnabled != value)
-                {
-                    _simulatedPressureEnabled = value;
-                    OnPropertyChanged();
-                    SchedulePersistSimulatedPressure();
-                }
-            }
-        }
-
-        public double SimulatedPressureStartTaperMm
-        {
-            get => _simulatedPressureStartTaperMm;
-            set
-            {
-                if (Math.Abs(_simulatedPressureStartTaperMm - value) > 0.0001)
-                {
-                    _simulatedPressureStartTaperMm = value;
-                    OnPropertyChanged();
-                    SchedulePersistSimulatedPressure();
-                }
-            }
-        }
-
-        public double SimulatedPressureEndTaperMm
-        {
-            get => _simulatedPressureEndTaperMm;
-            set
-            {
-                if (Math.Abs(_simulatedPressureEndTaperMm - value) > 0.0001)
-                {
-                    _simulatedPressureEndTaperMm = value;
-                    OnPropertyChanged();
-                    SchedulePersistSimulatedPressure();
-                }
-            }
-        }
-
-        public double SimulatedPressureSpeedMinMmPerSec
-        {
-            get => _simulatedPressureSpeedMinMmPerSec;
-            set
-            {
-                if (Math.Abs(_simulatedPressureSpeedMinMmPerSec - value) > 0.0001)
-                {
-                    _simulatedPressureSpeedMinMmPerSec = value;
-                    OnPropertyChanged();
-                    SchedulePersistSimulatedPressure();
-                }
-            }
-        }
-
-        public double SimulatedPressureSpeedMaxMmPerSec
-        {
-            get => _simulatedPressureSpeedMaxMmPerSec;
-            set
-            {
-                if (Math.Abs(_simulatedPressureSpeedMaxMmPerSec - value) > 0.0001)
-                {
-                    _simulatedPressureSpeedMaxMmPerSec = value;
-                    OnPropertyChanged();
-                    SchedulePersistSimulatedPressure();
-                }
-            }
-        }
-
-        public double SimulatedPressureFastSpeedMinFactor
-        {
-            get => _simulatedPressureFastSpeedMinFactor;
-            set
-            {
-                if (Math.Abs(_simulatedPressureFastSpeedMinFactor - value) > 0.0001)
-                {
-                    _simulatedPressureFastSpeedMinFactor = value;
-                    OnPropertyChanged();
-                    SchedulePersistSimulatedPressure();
-                }
-            }
-        }
-
-        public double SimulatedPressureFloor
-        {
-            get => _simulatedPressureFloor;
-            set
-            {
-                if (Math.Abs(_simulatedPressureFloor - value) > 0.0001)
-                {
-                    _simulatedPressureFloor = value;
-                    OnPropertyChanged();
-                    SchedulePersistSimulatedPressure();
-                }
-            }
-        }
-
-        public double SimulatedPressureEndFloor
-        {
-            get => _simulatedPressureEndFloor;
-            set
-            {
-                if (Math.Abs(_simulatedPressureEndFloor - value) > 0.0001)
-                {
-                    _simulatedPressureEndFloor = value;
-                    OnPropertyChanged();
-                    SchedulePersistSimulatedPressure();
-                }
-            }
-        }
-
-        public double SimulatedPressureSmoothingTauMs
-        {
-            get => _simulatedPressureSmoothingTauMs;
-            set
-            {
-                if (Math.Abs(_simulatedPressureSmoothingTauMs - value) > 0.0001)
-                {
-                    _simulatedPressureSmoothingTauMs = value;
-                    OnPropertyChanged();
-                    SchedulePersistSimulatedPressure();
-                }
-            }
-        }
-
-        private void ApplySimulatedPressureFields(SimulatedPressureConfig cfg)
-        {
-            _isApplyingSimulatedPressureFields = true;
-            try
-            {
-                _simulatedPressureEnabled = cfg.Enabled;
-                _simulatedPressureStartTaperMm = cfg.StartTaperMm;
-                _simulatedPressureEndTaperMm = cfg.EndTaperMm;
-                _simulatedPressureSpeedMinMmPerSec = cfg.SpeedMinMmPerSec;
-                _simulatedPressureSpeedMaxMmPerSec = cfg.SpeedMaxMmPerSec;
-                _simulatedPressureFastSpeedMinFactor = cfg.FastSpeedMinFactor;
-                _simulatedPressureFloor = cfg.PressureFloor;
-                _simulatedPressureEndFloor = cfg.EndPressureFloor;
-                _simulatedPressureSmoothingTauMs = cfg.SmoothingTauMs;
-
-                OnPropertyChanged(nameof(SimulatedPressureEnabled));
-                OnPropertyChanged(nameof(SimulatedPressureStartTaperMm));
-                OnPropertyChanged(nameof(SimulatedPressureEndTaperMm));
-                OnPropertyChanged(nameof(SimulatedPressureSpeedMinMmPerSec));
-                OnPropertyChanged(nameof(SimulatedPressureSpeedMaxMmPerSec));
-                OnPropertyChanged(nameof(SimulatedPressureFastSpeedMinFactor));
-                OnPropertyChanged(nameof(SimulatedPressureFloor));
-                OnPropertyChanged(nameof(SimulatedPressureEndFloor));
-                OnPropertyChanged(nameof(SimulatedPressureSmoothingTauMs));
-            }
-            finally
-            {
-                _isApplyingSimulatedPressureFields = false;
-            }
-        }
-
-        private SimulatedPressureConfig BuildSimulatedPressureConfigFromFields()
-        {
-            return new SimulatedPressureConfig
-            {
-                Enabled = _simulatedPressureEnabled,
-                StartTaperMm = _simulatedPressureStartTaperMm,
-                EndTaperMm = _simulatedPressureEndTaperMm,
-                SpeedMinMmPerSec = _simulatedPressureSpeedMinMmPerSec,
-                SpeedMaxMmPerSec = _simulatedPressureSpeedMaxMmPerSec,
-                FastSpeedMinFactor = _simulatedPressureFastSpeedMinFactor,
-                PressureFloor = (float)_simulatedPressureFloor,
-                EndPressureFloor = (float)_simulatedPressureEndFloor,
-                SmoothingTauMs = _simulatedPressureSmoothingTauMs
-            };
-        }
-
-        private void SchedulePersistSimulatedPressure()
-        {
-            if (_isApplyingSimulatedPressureFields) return;
-
-            _simulatedPressurePersistTimer ??= new DispatcherTimer(DispatcherPriority.Background, Dispatcher)
-            {
-                Interval = TimeSpan.FromMilliseconds(160)
-            };
-
-            _simulatedPressurePersistTimer.Stop();
-            _simulatedPressurePersistTimer.Tick -= SimulatedPressurePersistTimer_Tick;
-            _simulatedPressurePersistTimer.Tick += SimulatedPressurePersistTimer_Tick;
-            _simulatedPressurePersistTimer.Start();
-        }
-
-        private void SimulatedPressurePersistTimer_Tick(object? sender, EventArgs e)
-        {
-            _simulatedPressurePersistTimer?.Stop();
-            PersistSimulatedPressureNow();
-        }
-
-        private void PersistSimulatedPressureNow()
-        {
-            try
-            {
-                var cfg = BuildSimulatedPressureConfigFromFields();
-                cfg.ClampInPlace();
-                SettingsService.Instance.SetSimulatedPressureConfig(cfg);
-                ApplySimulatedPressureFields(cfg);
-            }
-            catch (Exception ex)
-            {
-                Debug.WriteLine($"[Settings] Failed to persist simulated pressure config: {ex}");
-            }
-        }
-
-        private void CleanupSimulatedPressureTimer(bool persistPending)
-        {
-            if (_simulatedPressurePersistTimer != null)
-            {
-                _simulatedPressurePersistTimer.Stop();
-                _simulatedPressurePersistTimer.Tick -= SimulatedPressurePersistTimer_Tick;
-                _simulatedPressurePersistTimer = null;
-            }
-
-            if (persistPending)
-            {
-                PersistSimulatedPressureNow();
-            }
-        }
-
         public SettingsWindow()
         {
             InitializeComponent();
@@ -536,221 +81,7 @@ namespace WindBoard
             RefreshCamouflagePreview(buildCache: false);
         }
 
-        private void BtnResetSimulatedPressure_Click(object sender, RoutedEventArgs e)
-        {
-            var cfg = DefaultSimulatedPressureConfig.Clone();
-            cfg.ClampInPlace();
-            ApplySimulatedPressureFields(cfg);
-            PersistSimulatedPressureNow();
-        }
-
-        private void PresetColor_Click(object sender, RoutedEventArgs e)
-        {
-            if (sender is Button btn && btn.Tag is string colorCode)
-            {
-                try
-                {
-                    Color color = (Color)ColorConverter.ConvertFromString(colorCode);
-                    CurrentColor = color;
-                    if (_colorPopupBox != null)
-                    {
-                        _colorPopupBox.IsPopupOpen = false;
-                    }
-                }
-                catch (Exception)
-                {
-                    // 无效颜色时默认使用白色
-                    CurrentColor = Colors.White;
-                    if (_colorPopupBox != null)
-                    {
-                        _colorPopupBox.IsPopupOpen = false;
-                    }
-                }
-            }
-        }
-
-        private async void OpenColorPicker_Click(object sender, RoutedEventArgs e)
-        {
-            if (_colorPopupBox != null)
-            {
-                _colorPopupBox.IsPopupOpen = false;
-            }
-
-            var colorPicker = new ColorPicker
-            {
-                Color = CurrentColor,
-                Width = 500,
-                Height = 300,
-                Margin = new Thickness(0, 0, 0, 16)
-            };
-
-            var stackPanel = new StackPanel { Margin = new Thickness(24) };
-            
-            var title = new TextBlock 
-            { 
-                Text = "自定义颜色", 
-                Style = (Style)FindResource("MaterialDesignHeadline6TextBlock"), 
-                Margin = new Thickness(0, 0, 0, 16) 
-            };
-            
-            var buttonPanel = new StackPanel 
-            { 
-                Orientation = Orientation.Horizontal, 
-                HorizontalAlignment = HorizontalAlignment.Right 
-            };
-            
-            var cancelButton = new Button 
-            { 
-                Content = "取消", 
-                Style = (Style)FindResource("MaterialDesignFlatButton"), 
-                IsCancel = true, 
-                Command = DialogHost.CloseDialogCommand, 
-                CommandParameter = false 
-            };
-            
-            var okButton = new Button 
-            { 
-                Content = "确定", 
-                Style = (Style)FindResource("MaterialDesignFlatButton"), 
-                IsDefault = true, 
-                Command = DialogHost.CloseDialogCommand, 
-                CommandParameter = true 
-            };
-
-            buttonPanel.Children.Add(cancelButton);
-            buttonPanel.Children.Add(okButton);
-
-            stackPanel.Children.Add(title);
-            stackPanel.Children.Add(colorPicker);
-            stackPanel.Children.Add(buttonPanel);
-
-            var result = await DialogHost.Show(stackPanel, "SettingsDialogHost");
-
-            if (result is bool confirmed && confirmed)
-            {
-                CurrentColor = colorPicker.Color;
-            }
-        }
-
-        private void ApplyHexColorFromTextBox()
-        {
-            var raw = HexTextBox?.Text;
-
-            var hex = raw?.Trim();
-            if (string.IsNullOrWhiteSpace(hex))
-            {
-                return;
-            }
-
-            if (!hex.StartsWith("#"))
-                hex = "#" + hex;
-
-            try
-            {
-                var color = (Color)ColorConverter.ConvertFromString(hex);
-                CurrentColor = color;
-            }
-            catch (Exception)
-            {
-                CurrentColor = Colors.White;
-            }
-        }
-
-        private void HexInput_KeyDown(object sender, KeyEventArgs e)
-        {
-            if (e.Key == Key.Enter)
-            {
-                ApplyHexColorFromTextBox();
-                e.Handled = true;
-            }
-        }
-
-        private void HexApply_Click(object sender, RoutedEventArgs e)
-        {
-            ApplyHexColorFromTextBox();
-        }
-
-        private void BtnResetColor_Click(object sender, RoutedEventArgs e)
-        {
-            if (sender is Button btn && btn.Tag is string colorCode)
-            {
-                try
-                {
-                    var color = (Color)ColorConverter.ConvertFromString(colorCode);
-                    CurrentColor = color;
-                }
-                catch (Exception)
-                {
-                    // 无效颜色时默认使用白色
-                    CurrentColor = Colors.White;
-                }
-            }
-        }
-
-        private void CloseButton_Click(object sender, RoutedEventArgs e)
-        {
-            Close();
-        }
-
-        // --- 基本设置·视频展台：事件处理 ---
-        private void ToggleVideoPresenter_Checked(object sender, RoutedEventArgs e)
-        {
-            VideoPresenterEnabled = true;
-        }
-
-        private void ToggleVideoPresenter_Unchecked(object sender, RoutedEventArgs e)
-        {
-            VideoPresenterEnabled = false;
-        }
-
-        private void BtnBrowseVideoPresenterPath_Click(object sender, RoutedEventArgs e)
-        {
-            try
-            {
-                var dlg = new OpenFileDialog
-                {
-                    Title = "选择视频展台程序",
-                    Filter = "可执行文件 (*.exe)|*.exe|所有文件 (*.*)|*.*",
-                    CheckFileExists = true
-                };
-                if (!string.IsNullOrWhiteSpace(VideoPresenterPath))
-                {
-                    try
-                    {
-                        dlg.InitialDirectory = Path.GetDirectoryName(VideoPresenterPath) ?? Environment.GetFolderPath(Environment.SpecialFolder.ProgramFilesX86);
-                    }
-                    catch { }
-                }
-                if (dlg.ShowDialog(this) == true)
-                {
-                    VideoPresenterPath = dlg.FileName;
-                    try { SettingsService.Instance.SetVideoPresenterPath(VideoPresenterPath); } catch { }
-                }
-            }
-            catch { }
-        }
-
-        private void BtnApplyVideoPresenterPath_Click(object sender, RoutedEventArgs e)
-        {
-            try { SettingsService.Instance.SetVideoPresenterPath(VideoPresenterPath); } catch { }
-        }
-
-        private void BtnApplyVideoPresenterArgs_Click(object sender, RoutedEventArgs e)
-        {
-            try { SettingsService.Instance.SetVideoPresenterArgs(VideoPresenterArgs); } catch { }
-        }
-
-        private void BtnResetVideoPresenterDefaults_Click(object sender, RoutedEventArgs e)
-        {
-            VideoPresenterPath = DefaultVideoPresenterPath;
-            VideoPresenterArgs = DefaultVideoPresenterArgs;
-            try
-            {
-                SettingsService.Instance.SetVideoPresenterPath(VideoPresenterPath);
-                SettingsService.Instance.SetVideoPresenterArgs(VideoPresenterArgs);
-            }
-            catch { }
-        }
+        private void CloseButton_Click(object sender, RoutedEventArgs e) => Close();
 
         // 统一底部按钮：应用/确定/取消
         private void ApplyAllSettings()
@@ -765,10 +96,7 @@ namespace WindBoard
             PersistSimulatedPressureNow();
         }
 
-        private void BtnApply_Click(object sender, RoutedEventArgs e)
-        {
-            ApplyAllSettings();
-        }
+        private void BtnApply_Click(object sender, RoutedEventArgs e) => ApplyAllSettings();
 
         private void BtnOk_Click(object sender, RoutedEventArgs e)
         {
@@ -776,122 +104,12 @@ namespace WindBoard
             Close();
         }
 
-        private void BtnCancel_Click(object sender, RoutedEventArgs e)
-        {
-            Close();
-        }
+        private void BtnCancel_Click(object sender, RoutedEventArgs e) => Close();
 
         protected override void OnClosed(EventArgs e)
         {
             CleanupSimulatedPressureTimer(persistPending: true);
             base.OnClosed(e);
-        }
-
-        // --- 基本设置·伪装 ---
-        private void ToggleCamouflage_Checked(object sender, RoutedEventArgs e)
-        {
-            CamouflageEnabled = true;
-        }
-
-        private void ToggleCamouflage_Unchecked(object sender, RoutedEventArgs e)
-        {
-            CamouflageEnabled = false;
-        }
-
-        private void BtnBrowseCamouflageIcon_Click(object sender, RoutedEventArgs e)
-        {
-            try
-            {
-                var dlg = new OpenFileDialog
-                {
-                    Title = "选择程序或图标",
-                    Filter = "可执行/图标/图片|*.exe;*.ico;*.png;*.jpg;*.jpeg;*.bmp|所有文件|*.*",
-                    CheckFileExists = true
-                };
-                if (!string.IsNullOrWhiteSpace(CamouflageSourcePath))
-                {
-                    try
-                    {
-                        dlg.InitialDirectory = Path.GetDirectoryName(CamouflageSourcePath) ?? Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments);
-                    }
-                    catch { }
-                }
-                if (dlg.ShowDialog(this) == true)
-                {
-                    CamouflageSourcePath = dlg.FileName;
-                }
-            }
-            catch
-            {
-                // ignore dialog errors
-            }
-        }
-
-        private void BtnClearCamouflageIcon_Click(object sender, RoutedEventArgs e)
-        {
-            CamouflageSourcePath = string.Empty;
-            CamouflageIconPreview = null;
-            CamouflageSourceDisplayName = "未选择文件";
-            try { SettingsService.Instance.SetCamouflageIconCachePath(string.Empty); } catch { }
-        }
-
-        private void RefreshCamouflagePreview(bool buildCache)
-        {
-            var sourcePath = CamouflageSourcePath;
-            if (string.IsNullOrWhiteSpace(sourcePath) || !File.Exists(sourcePath))
-            {
-                CamouflageIconPreview = null;
-                CamouflageSourceDisplayName = "未选择文件";
-                if (buildCache)
-                {
-                    try { SettingsService.Instance.SetCamouflageIconCachePath(string.Empty); } catch { }
-                }
-                return;
-            }
-
-            CamouflageSourceDisplayName = Path.GetFileName(sourcePath);
-
-            if (buildCache)
-            {
-                if (CamouflageService.Instance.TryBuildIconCache(sourcePath, out var cachePath, out var preview))
-                {
-                    CamouflageIconPreview = preview;
-                    try { SettingsService.Instance.SetCamouflageIconCachePath(cachePath); } catch { }
-                }
-                else
-                {
-                    CamouflageIconPreview = null;
-                    CamouflageSourceDisplayName = "无法读取图标";
-                    try { SettingsService.Instance.SetCamouflageIconCachePath(string.Empty); } catch { }
-                }
-                return;
-            }
-
-            // 不重建缓存时，尝试从已有缓存加载预览
-            var cachedPath = SettingsService.Instance.GetCamouflageIconCachePath();
-            if (!string.IsNullOrWhiteSpace(cachedPath) && File.Exists(cachedPath))
-            {
-                try
-                {
-                    CamouflageIconPreview = CamouflageService.Instance.LoadIconFromFile(cachedPath);
-                }
-                catch
-                {
-                    CamouflageIconPreview = null;
-                }
-            }
-            else
-            {
-                // 尝试直接用源文件作为预览（不写缓存）
-                try
-                {
-                    CamouflageIconPreview = CamouflageService.Instance.LoadIconFromFile(sourcePath);
-                }
-                catch
-                {
-                    CamouflageIconPreview = null;
-                }
-            }
         }
     }
 }
